requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
dnspython==2.4.2
idna==3.6
python-whois==0.8.0
requests==2.31.0
urllib3==2.1.0
